Refactor: Separate SetupScreen into SetupCamera and SetupServer
This commit refactors the `SetupScreen` into two distinct composables: `SetupCamera` and `SetupServer`. This separation improves code organization and modularity.

**Key Changes:**

- **Deleted `SetupCamera.kt` (old version).**
- **Created `SetupCamera.kt` (new version):**
    - This composable now handles camera permission requests.
    - It uses `rememberPermissionState` from `com.google.accompanist.permissions`.
    - It displays messages and buttons based on the permission status (granted, rationale needed, denied).
    - It allows the user to proceed without camera permission.
    - Layout adapts to portrait and landscape orientations.
- **Created `SetupServer.kt`:**
    - This composable handles the server configuration (IP address and port).
    - It reuses the existing UI logic for input fields, connection testing, and saving settings.
    - It uses `SetupViewModel` for its logic.
- **Created `SetupCameraViewModel.kt`:**
    - A new ViewModel to manage the state for `SetupCamera` (e.g., `canContinue`, `iWasRationale`, `permanentDenied`).
- **Updated `SetupScreen.kt`:**
    - Now uses a `HorizontalPager` to display `SetupCamera` and `SetupServer` as separate pages.
    - The initial page depends on whether camera permission is already granted.
    - User scrolling between pages is disabled; navigation is programmatic.
    - `onSetupCompleted` callback is passed down to `SetupServer`.
    - `SetupCamera` navigates to `SetupServer` upon its completion (permission granted or skipped).

**Overall:**
The setup process is now a two-step flow within a `HorizontalPager`: first camera permission, then server configuration. This change makes the setup UI more guided and separates concerns effectively.
